pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shms-calicration-nn"
version = "0.1.0"
description = "Extracted SHMS optics neural-network training toolkit"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "LantoZheng"}
]
dependencies = [
  "numpy>=1.23",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "pyyaml>=6.0",
  "matplotlib>=3.7",
  "seaborn>=0.12",
  "torch>=2.0",
  "uproot>=5.0",
  "awkward>=2.0",
  "onnx>=1.14",
  "onnxruntime>=1.16",
]

[tool.pytest.ini_options]
testpaths = ["training/tests"]
